use css3d to replace the active hand indicator with a rotating pyramid
