+++ packages/frontend-react/src/components/Hand/ActiveHandIndicator/index.ts
@@ -1,0 +1,1 @@
+export { ActiveHandIndicator } from './ActiveHandIndicator';
