--- packages/frontend-react/src/components/Hand/Hand.module.css
@@ -1,45 +1,0 @@
-/* TODO: Make this triangle icon better, gradient BG etc. */
-.active {
-  position: absolute;
-  top: 2px;
-
-  width: 0;
-  height: 0;
-  border-left: 10px solid transparent;
-  border-right: 10px solid transparent;
-  border-top: 10px solid yellow;
-
-  animation: roty 4s linear infinite;
-}
-
-.active[data-direction='left'] {
-  left: 39px;
-}
-
-.active[data-direction='right'] {
-  right: 39px;
-}
-
-.score {
-  z-index: 99;
-  position: absolute;
-  color: white;
-  bottom: 5px;
-  font-size: 15px;
-  font-family: 'Press Start 2P';
-  text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
-}
-
-.score[data-direction='left'] {
-  left: 42px;
-}
-
-.score[data-direction='right'] {
-  right: 42px;
-}
-
-@keyframes roty {
-  100% {
-    transform: rotate3d(0, 1, 0, 360deg);
-  }
-}
