+++ packages/frontend-react/src/components/Hand/Hand.tsx
@@ -2,38 +2,8 @@ import * as TripleTriad from '@tripletriad/game';
 import clsx from 'clsx';
 import React from 'react';
 import { Card } from '../Card';
-import styles from './Hand.module.css';
-
-type ActiveHandIndicatorProps = {
-  active: boolean;
-  player: TripleTriad.Player;
-};
-
-const ActiveHandIndicator = ({ active, player }: ActiveHandIndicatorProps) => {
-  return (
-    <span
-      className={clsx('m-auto', active ? styles.active : null)}
-      hidden={active ? false : true}
-      data-direction={player.label == TripleTriad.Player.One ? 'left' : 'right'}
-    />
-  );
-};
-
-type ScoreProps = {
-  score: number;
-  player: TripleTriad.Player;
-};
-
-const Score = ({ score, player }: ScoreProps) => {
-  return (
-    <span
-      className={clsx('m-auto', styles.score)}
-      data-direction={player.label == TripleTriad.Player.One ? 'left' : 'right'}
-    >
-      {score}
-    </span>
-  );
-};
+import { ActiveHandIndicator } from './ActiveHandIndicator';
+import { Score } from './Score';
 
 type HandProps = {
   player: TripleTriad.Player;
