+++ packages/frontend-react/src/components/Hand/Score/Score.module.css
@@ -1,0 +1,17 @@
+.score {
+  z-index: 99;
+  position: absolute;
+  color: white;
+  bottom: 5px;
+  font-size: 15px;
+  font-family: 'Press Start 2P';
+  text-shadow: -1px 0 black, 0 1px black, 1px 0 black, 0 -1px black;
+}
+
+.score[data-direction='left'] {
+  left: 42px;
+}
+
+.score[data-direction='right'] {
+  right: 42px;
+}
