+++ packages/frontend-react/src/components/Hand/Score/index.ts
@@ -1,0 +1,1 @@
+export { Score } from './Score';
